requirements.txt
agent-framework-azure-ai --pre
azure-identity>=1.17.1
python-dotenv>=1.0.1
pydantic>=2.0.0
streamlit>=1.28.0